requirements.txt
numpy
flask==2.2.3
gunicorn==20.1.0
flask-cors==3.0.10
soundfile==0.12.1
google.cloud.texttospeech
